+++ run_trigger_integration_test.go
@@ -233,6 +233,35 @@ func TestRunTriggerRead(t *testing.T) {
 	})
 }
 
+func TestRunTriggerReadWithOptions(t *testing.T) {
+	client := testClient(t)
+	ctx := context.Background()
+
+	orgTest, orgTestCleanup := createOrganization(t, client)
+	defer orgTestCleanup()
+
+	wTest, wTestCleanup := createWorkspace(t, client, orgTest)
+	defer wTestCleanup()
+
+	sourceableTest, sourceableTestCleanup := createWorkspace(t, client, orgTest)
+	defer sourceableTestCleanup()
+
+	rtTest, rtTestCleanup := createRunTrigger(t, client, wTest, sourceableTest)
+	defer rtTestCleanup()
+
+	t.Run("with include options", func(t *testing.T) {
+		rt, err := client.RunTriggers.ReadWithOptions(ctx, rtTest.ID, &RunTriggerReadOptions{
+			Include: []RunTriggerIncludeOpt{RunTriggerSourceable, RunTriggerWorkspace},
+		})
+		require.NoError(t, err)
+		assert.Equal(t, rtTest.ID, rt.ID)
+		require.NotNil(t, rt.Sourceable)
+		assert.Equal(t, sourceableTest.ID, rt.Sourceable.ID)
+		require.NotNil(t, rt.Workspace)
+		assert.Equal(t, wTest.ID, rt.Workspace.ID)
+	})
+}
+
 func TestRunTriggerDelete(t *testing.T) {
 	t.Parallel()
 	client := testClient(t)